+++ NetStatsMonitor/Configuration/Graphs/GraphConfiguration.cs
@@ -15,15 +15,20 @@ namespace Unity.Multiplayer.Tools.NetStatsMonitor
         /// The number of samples that are maintained for the purpose of graphing.
         /// </summary>
         /// <remarks>
-        /// If the value is out of the range [0, 4096], it will be clamped to the
-        /// nearest value.
+        /// The value is clamped to the range [8, 4096].
         /// </remarks>
         [field: SerializeField]
         [field: Tooltip("The number of samples that are maintained for the purpose of graphing. " +
-                        "If the value is out of the range [0, 4096], it will be clamped to the nearest value.")]
+                        "The value is clamped to the range [8, 4096].")]
         [field: Range(ConfigurationLimits.k_GraphSampleMin, ConfigurationLimits.k_GraphSampleMax)]
         int m_SampleCount = 256;
 
+        /// <summary>
+        /// The number of samples that are maintained for the purpose of graphing.
+        /// </summary>
+        /// <remarks>
+        /// The value is clamped to the range [8, 4096].
+        /// </remarks>
         public int SampleCount
         {
             get => m_SampleCount;